Admin panel

+++ services/usermanagement/src/controllers/userController.js
@@ -1,6 +1,9 @@
 const UserAssessment = require("../models/assessmentModel");
 const User = require("../models/userModel");
 const getMonthName = require("../utils/monthName");
+const generateOtp = require("../utils/otpGenerator");
+const sendEmail = require("../utils/sendemail");
+
 // GET: Get all users from DB
 exports.getAllUsers = async (req, res) => {
   try {
@@ -43,7 +46,6 @@ exports.getAllUsers = async (req, res) => {
         },
       },
     ]);
-    console.log("TEST: ", test);
 
     res.status(200).json({ status: "Success", data });
   } catch (error) {
@@ -80,7 +82,6 @@ exports.createNewAdmin = async (req, res) => {
 
 // GET: Get user by ID
 exports.getUser = async (req, res) => {
-  console.log(req.params);
   try {
     const user = await User.findById(req.params.id).select("+createdAt");
     res.status(200).json({ status: "Success", user });
@@ -93,9 +94,17 @@ exports.getUser = async (req, res) => {
 
 // UPDATE: Update user details
 exports.updateUser = async (req, res) => {
-  const { id, password } = req.body;
-  const updatedUser = await User.findByIdAndUpdate(id, { password });
-  res.status(200).json({ status: "Success", message: "User Password Updated" });
+  try {
+    const { id, password } = req.body;
+    const updatedUser = await User.findByIdAndUpdate(id, { password });
+    res
+      .status(200)
+      .json({ status: "Success", message: "User Password Updated" });
+  } catch (error) {
+    res
+      .status(400)
+      .json({ status: "Failed", message: "User Password Not Updated" });
+  }
 };
 
 // DELETE: Delete user
@@ -106,12 +115,20 @@ exports.updateUser = async (req, res) => {
 exports.deleteUser = async (req, res) => {
   try {
     const userID = req.params.id;
-    await User.findByIdAndDelete(req.params.id);
-
+    const user = await User.findOne({ _id: userID }).select("+active");
+    // If user is already is inactive - then delete it from db - When admin requst to
+    if (!user.active) {
+      await User.findByIdAndDelete(req.params.id);
+    } else {
+      // make user inactive
+      await User.findByIdAndUpdate(userID, { $set: { active: false } });
+    }
     // Response
-    res
-      .status(200)
-      .json({ status: "Success", message: "User Deleted", data: null });
+    res.status(200).json({
+      status: "Success",
+      message: "User Deactivated | Deleted",
+      data: null,
+    });
   } catch (error) {
     res.status(200).json({ status: "Failed", error });
   }
@@ -123,7 +140,6 @@ exports.deleteUser = async (req, res) => {
 // 3) Access assessment for each user, referencing their user id in assessment document
 
 exports.storeUserSolarAssessment = async (req, res) => {
-  console.log("User Management: History Controller:", req.body);
   try {
     const {
       user,
@@ -154,7 +170,6 @@ exports.storeUserSolarAssessment = async (req, res) => {
       data: userAssessment,
     });
   } catch (error) {
-    console.error("Error creating UserAssessment:", error.message);
     res.status(400).json({
       status: "Failed",
       message: error.message,
@@ -222,5 +237,181 @@ exports.userTimeLineStats = async (req, res) => {
     res.status(200).json({ status: "Success", data });
   } catch (error) {
     res.status(400).json({ status: "Failed", error });
+  }
+};
+
+// Foget Password controller, Generate OTP, send 1 copy to user and store 1 copy in user document
+exports.forgetPassword = async (req, res) => {
+  try {
+    // get user email
+    const { email } = req.body;
+
+    const user = await User.findOne({ email });
+
+    /**
+     * verify user
+     * send otp
+     * Success response
+     */
+
+    // If no user instant fallback
+    if (!user)
+      return res
+        .status(404)
+        .json({ status: "Failed", message: "No user found with this email" });
+    // If user then send 1 copy of otp via email an store one copy to db
+    if (user) {
+      // Get 4 digit otp: Utility function
+      const otp = generateOtp();
+
+      // Send one copy to user via email
+      sendEmail(
+        process.env.EMAIL_SENDER,
+        process.env.ADMIN_EMAIL,
+        email, // user email which will recieve email
+        "Forget Password",
+        `<b>This is your forget password otp</b>.
+        <p>OTP <b>${otp}</b></p>
+        <p>Verify your otp and reset your password</p>
+        <p><i>If you did not request this otp, you can successfully ignore this message.</i></p>`
+      ).catch((error) => console.log("OTP Email Sending Error: ", error));
+
+      /**
+       * AS WE HAVE USER, NOW WE WILL UPDATE USER BY ID
+       */
+      // Store one copy in db
+      // TODO: Fix Bug Use controller forgetPassword: This lines does not update the specified user, it update the first document only check it and find a fix for it. (It is commented)
+      // await User.findOneAndUpdate({ email, $set: { otp } });
+      await User.findByIdAndUpdate(user._id, { $set: { otp } });
+      // Success response
+      res.status(200).json({
+        status: "Success",
+        message:
+          "OTP successfully sent. Please check your email and verify otp.",
+      });
+    }
+  } catch (error) {
+    // Error response
+    res.status(400).json({ status: "Failed", error });
+  }
+};
+
+/**
+ * This controller check user, then check user otp and then update otpVerified filed to true
+ * @param {*} req
+ * @param {*} res
+ */
+exports.otpVerification = async (req, res) => {
+  try {
+    const { email, otp } = req.body;
+
+    /**
+     * Check user by email,
+     * Then verify otp,
+     * If otp is valid then update the otpVerified filed to true.
+     * */
+
+    // check user
+    const user = await User.findOne({ email });
+    // If no user then instant fallback
+    if (!user) {
+      return res
+        .status(404)
+        .json({ status: "Failed", message: "No user with this email" });
+    }
+
+    // If User then verify user otp and incase of invalid fallback
+    if (user.otp !== otp) {
+      return res
+        .status(400)
+        .json({ status: "Failed", message: "Invaid OTP. Provide Valid OTP" });
+    }
+
+    // TODO: Fix Bug Use controller forgetPassword: This lines does not update the specified user, it update the first document only check it and find a fix for it. (It is commented)
+    // await User.findOneAndUpdate({ email, $set: { otpVerified: true } });
+
+    // If User reached at this point. Then the otp is valid, update the otpVerified field
+    await User.findByIdAndUpdate(user._id, { $set: { otpVerified: true } });
+    // Success response
+    res.status(200).json({ status: "Success", message: "OTP is verified." });
+  } catch (error) {
+    // Error response
+    res.status(400).status({ status: "Failed", error });
+  }
+};
+
+/**
+ * This update user password. Check User - Check OTP Verification - Check Passwords then Update password
+ * @param {*} req
+ * @param {*} res
+ */
+exports.resetPassword = async (req, res) => {
+  try {
+    const { email, password, confirmPassword } = req.body;
+
+    // Get user
+    const user = await User.findOne({ email });
+
+    // Instant Fallback incase of invalid email
+    if (!user)
+      return res.status(404).json({
+        status: "Failed",
+        message: "These is no user with this email",
+      });
+
+    if (user) {
+      // Check OTP Verification
+      if (!user.otpVerified)
+        return res.status(400).json({
+          status: "Failed",
+          message: "Invlid OTP. Please verify your otp.",
+        });
+
+      // Confirm passwrod
+      if (password !== confirmPassword)
+        return res
+          .status(402)
+          .json({ status: "Failed", message: "Both password should match" });
+
+      // TODO: Fix Bug Use controller forgetPassword: This lines does not update the specified user, it update the first document only check it and find a fix for it. (It is commented)
+      // If user reaced at this point then every thing is ok, update the password
+      // Update otpVerified to false for next time passwrod reset verification
+      // await User.findOneAndUpdate({
+      //   email,
+      //   $set: { password, otpVerified: false, otp: null },
+      // });
+
+      await User.findByIdAndUpdate(user._id, {
+        $set: { password, otpVerified: false, otp: null },
+      });
+      // Success response
+      res
+        .status(200)
+        .json({ status: "Success", message: "Your password has been reset." });
+    }
+  } catch (error) {
+    // Error Response
+    res.status(400).json({ status: "Failed", error });
+  }
+};
+
+// Send public feedback
+exports.sendFeedback = async (req, res) => {
+  try {
+    const { name, email, subject, message } = req.body;
+    sendEmail(
+      process.env.FEEDBACK_NAME,
+      process.env.ADMIN_EMAIL,
+      process.env.FEEDBACK_EMAIL, // feedback email which revice feedback (admin email)
+      subject,
+      `<b>This is a Feeback Email from ${name}.</b>.
+      <p>User Email is ${email}</p>
+      <p>User Message is: ${message}</p>`
+    ).catch((error) => console.log("Feedback Email Sending Error: ", error));
+    res.status(200).json({ status: "Success", message: "Feedback Recieved" });
+  } catch (error) {
+    res
+      .status(400)
+      .json({ status: "Failed", message: "Feedback Not sent", error });
   }
 };
